Delete Local Guide functionality added
A new algorithm object AlgorithmContainer is added to facilitate sorting
in both DeleteLocalGuideActivity as well as DisplayLocalGuidesActivity.
DeleteLocalGuideActivity displays all the guides in the form of button
just like DisplayLocalGuidesActivity but upon pressing the button it
prompts the user whether they wish to delete or not and pressing yes in
this prompt does so.